+++ src/components/Button/Button.stories.tsx
@@ -52,8 +52,26 @@ export const big = (): React.ReactElement => (
   </Button>
 )
 
+export const small = (): React.ReactElement => (
+  <Button type="button" small>
+    Click Me
+  </Button>
+)
+
+export const icon = (): React.ReactElement => (
+  <Button type="button" icon>
+    Click Me
+  </Button>
+)
+
 export const unstyled = (): React.ReactElement => (
   <Button type="button" unstyled>
+    Click Me
+  </Button>
+)
+
+export const customClass = (): React.ReactElement => (
+  <Button type="button" className="custom-class">
     Click Me
   </Button>
 )