+++ src/components/Table/Table.module.css.d.ts
@@ -4,5 +4,5 @@ interface CssExports {
   fixed: string
   fullwidth: string
 }
-declare const cssExports: CssExports
-export = cssExports
+export const cssExports: CssExports
+export default cssExports