+++ src/components/forms/Checkbox/Checkbox.test.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import { render } from '@testing-library/react'
+
+import { Checkbox } from './Checkbox'
+
+describe('Checkbox component', () => {
+  it('renders without errors', () => {
+    const { queryByTestId } = render(
+      <Checkbox id="input-checkbox" name="input-checkbox" label="My checkbox" />
+    )
+    expect(queryByTestId('checkbox')).toBeInTheDocument()
+  })
+})