+++ src/components/forms/ErrorMessage/ErrorMessage.stories.tsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import { ErrorMessage } from './ErrorMessage'
+
+export default {
+  title: 'ErrorMessage',
+  parameters: {
+    info: `
+USWDS 2.0 ErrorMessage component
+
+Source: https://designsystem.digital.gov/components/form-controls/
+`,
+  },
+}
+
+export const defaultErrorMessage = (): React.ReactElement => (
+  <ErrorMessage>Helpful error message</ErrorMessage>
+)