+++ src/components/forms/TextInput/TextInput.test.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import { render } from '@testing-library/react'
+
+import { TextInput } from './TextInput'
+
+describe('TextInput component', () => {
+  it('renders without errors', () => {
+    const { queryByTestId } = render(
+      <TextInput id="input-type-text" name="input-type-text" type="text" />
+    )
+    expect(queryByTestId('textInput')).toBeInTheDocument()
+  })
+})