+++ src/index.ts
@@ -1,3 +1,4 @@
+/** USWDS basic components */
 export { Alert } from './components/Alert/Alert'
 export { Accordion } from './components/Accordion/Accordion'
 export { GovBanner } from './components/GovBanner/GovBanner'
@@ -5,3 +6,26 @@ export { Button } from './components/Button/Button'
 export { Table } from './components/Table/Table'
 export { Tag } from './components/Tag/Tag'
 export { SideNav } from './components/SideNav/SideNav'
+
+/** Form components */
+export { Checkbox } from './components/forms/Checkbox/Checkbox'
+export { Dropdown } from './components/forms/Dropdown/Dropdown'
+export { ErrorMessage } from './components/forms/ErrorMessage/ErrorMessage'
+export { Fieldset } from './components/forms/Fieldset/Fieldset'
+export { Form } from './components/forms/Form/Form'
+export { FormGroup } from './components/forms/FormGroup/FormGroup'
+export { Label } from './components/forms/Label/Label'
+export { Radio } from './components/forms/Radio/Radio'
+export { Textarea } from './components/forms/Textarea/Textarea'
+export { TextInput } from './components/forms/TextInput/TextInput'
+
+/** Truss-designed components */
+export {
+  Modal,
+  Overlay,
+  ModalContainer,
+  connectModal,
+  ConnectedModalProps,
+  useModal,
+  ModalHook,
+} from './components/Modal/Modal'