+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2020-03-19
+
+- Add initial set of USWDS form components (Checkbox, Dropdown, ErrorMessage, Fieldset, Form, FormGroup, Label, Radio, TextInput, Textarea)
+- Add native HTML attributes to existing Tag & Button components in order to support custom CSS classes
+- Added new Modal component (designed by Truss) and related Overlay and ModalContainer components
+- Configured use of SASS/SCSS and SASS/SCSS modules so that we can use USWDS SCSS
+- Configure Dependabot for automated dependency updates
+
 ## [1.0.0] - 2019-01-10
 
 ### Added